pyproject.toml
[tool.poetry]
name = "ids"
version = "0.1.0"
description = "Intelligent Development System - Multi-agent deliberation platform"
authors = ["Oleksiy"]
readme = "README.md"
packages = [{include = "ids"}]

[tool.poetry.dependencies]
python = "^3.11"
# AI/LLM
google-generativeai = "^0.3.0"
anthropic = "^0.40.0"
# Telegram
python-telegram-bot = "^21.0"
# Storage
motor = "^3.3.2"
pymongo = "^4.6.0"
chromadb = "^0.5.0"
redis = "^5.0.0"
# Data validation
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
# Utilities
python-dotenv = "^1.0.0"
structlog = "^24.1.0"
pyyaml = "^6.0.1"
# Async
aiohttp = "^3.9.0"
# NumPy version constraint for compatibility
numpy = "^1.24.0,<2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.13"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
